(WIP) Unittest for _make_mpl_fig() created
Unittest for _make_mpl_fig() written. Uses a stock dataset stored in
saved_sample_data.json to generate a known image using _make_mpl_fig().
A histogram of the output is generated and compared against a recorded
histogram using the Kolmogorov–Smirnov test. The images are considered
identical if the p-value for the test is >= 0.95.
Test should be cleaned up and moved to a different TestCase class.
Language-support needs to be re-enabled and a new histogram needs to be
generated for an es_MX image-output.